+++ test/src/kava-event-model.spec.js
@@ -48,6 +48,58 @@ class FakeModel {
   getAverageBitrate() {
     return 600;
   }
+
+  getBandwidth() {
+    return 3000;
+  }
+
+  getMaxManifestDownloadTime() {
+    return 100;
+  }
+
+  getSoundMode() {
+    return 1;
+  }
+
+  getTabMode() {
+    return 1;
+  }
+
+  getAvailableBuffer() {
+    return 2;
+  }
+
+  getDroppedFramesRatio() {
+    return 0.01;
+  }
+
+  getSegmentDownloadTime() {
+    return 0.3;
+  }
+
+  getPlayerJSLoadTime() {
+    return 0.23;
+  }
+
+  getForwardBufferHealth() {
+    return 0.9;
+  }
+
+  getTargetBuffer() {
+    return 30;
+  }
+
+  getNetworkConnectionType() {
+    return '4g';
+  }
+
+  getNetworkConnectionOverhead() {
+    return '0.12';
+  }
+
+  getFlavorParamsId() {
+    return 32;
+  }
 }
 
 describe('KavaEventModel', () => {
@@ -63,14 +115,27 @@ describe('KavaEventModel', () => {
       actualBitrate: fakeModel.getActualBitrate(),
       averageBitrate: fakeModel.getAverageBitrate(),
       captionsLanguage: fakeModel.getCaption(),
-      audioLanguage: fakeModel.getLanguage()
+      audioLanguage: fakeModel.getLanguage(),
+      bandwidth: fakeModel.getBandwidth(),
+      droppedFramesRatio: fakeModel.getDroppedFramesRatio(),
+      manifestDownloadTime: fakeModel.getMaxManifestDownloadTime(),
+      soundMode: fakeModel.getSoundMode(),
+      tabMode: fakeModel.getTabMode(),
+      segmentDownloadTime: fakeModel.getSegmentDownloadTime(),
+      forwardBufferHealth: fakeModel.getForwardBufferHealth(),
+      targetBuffer: fakeModel.getTargetBuffer(),
+      networkConnectionType: fakeModel.getNetworkConnectionType(),
+      networkConnectionOverhead: fakeModel.getNetworkConnectionOverhead(),
+      flavorParamsId: fakeModel.getFlavorParamsId()
     });
   });
 
   it('IMPRESSION', () => {
     KavaEventModel.IMPRESSION.type.should.equal('IMPRESSION');
     KavaEventModel.IMPRESSION.index.should.equal(1);
-    KavaEventModel.IMPRESSION.getEventModel(fakeModel).should.deep.equal({});
+    KavaEventModel.IMPRESSION.getEventModel(fakeModel).should.deep.equal({
+      playerJSLoadTime: fakeModel.getPlayerJSLoadTime()
+    });
   });
 
   it('PLAY_REQUEST', () => {
@@ -87,7 +152,8 @@ describe('KavaEventModel', () => {
       bufferTimeSum: fakeModel.getBufferTimeSum(),
       actualBitrate: fakeModel.getActualBitrate(),
       joinTime: fakeModel.getJoinTime(),
-      canPlay: fakeModel.getCanPlayTime()
+      canPlay: fakeModel.getCanPlayTime(),
+      networkConnectionType: fakeModel.getNetworkConnectionType()
     });
   });